추가작업

+++ src/main/java/com/ll/exam/final__2022_10_08/app/rebate/service/RebateService.java
@@ -53,6 +53,10 @@ public class RebateService {
         RebateOrderItem oldRebateOrderItem = rebateOrderItemRepository.findByOrderItemId(item.getOrderItem().getId()).orElse(null);
 
         if (oldRebateOrderItem != null) {
+            if ( oldRebateOrderItem.isRebateDone() ) {
+                return;
+            }
+
             rebateOrderItemRepository.delete(oldRebateOrderItem);
         }
